-김동욱- widget에 폴더들 보기 불편하대서 widget 바로 밑으로 다 뺐습니다.
