+++ lib/Widget/alarm/alarm1.dart
@@ -1,22 +1,22 @@
-import 'package:dating/Widget/main/alarm/alarm_widget.dart';
+import 'package:dating/Widget/alarm/alarm_widget.dart';
 import 'package:flutter/material.dart';
 
-class Alarm10 extends StatelessWidget {
-  const Alarm10({super.key});
+class Alarm1 extends StatelessWidget {
+  const Alarm1({super.key});
 
   @override
   Widget build(BuildContext context) {
     return const AlarmWidget(
       text: Text(
-        '남동길',
+        '홍길동',
         style: TextStyle(fontSize: 17),
       ),
       text2: Text(
-        '나에게 채팅을 보냈네요?"안녕하세요? 뭐하고 계세 \n요?"',
+        '나에게 하트를 눌렀어요! 마음에 드시나요?',
         style: TextStyle(fontSize: 12, color: Colors.grey),
       ),
       text3: Text(
-        '10월 17일',
+        '1시간전',
         style: TextStyle(color: Colors.grey, fontSize: 13),
       ),
     );
