+++ lib/Widget/alarm/alarm8.dart
@@ -1,8 +1,8 @@
-import 'package:dating/Widget/main/alarm/alarm_widget.dart';
+import 'package:dating/Widget/alarm/alarm_widget.dart';
 import 'package:flutter/material.dart';
 
-class Alarm4 extends StatelessWidget {
-  const Alarm4({super.key});
+class Alarm8 extends StatelessWidget {
+  const Alarm8({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -16,7 +16,7 @@ class Alarm4 extends StatelessWidget {
         style: TextStyle(fontSize: 12, color: Colors.grey),
       ),
       text3: Text(
-        '10월 13일',
+        '10월 14일',
         style: TextStyle(color: Colors.grey, fontSize: 13),
       ),
     );
