+++ lib/screen/main/alarm_screen.dart
@@ -1,13 +1,13 @@
-import 'package:dating/Widget/main/alarm/alarm1.dart';
-import 'package:dating/Widget/main/alarm/alarm10.dart';
-import 'package:dating/Widget/main/alarm/alarm2.dart';
-import 'package:dating/Widget/main/alarm/alarm3.dart';
-import 'package:dating/Widget/main/alarm/alarm4.dart';
-import 'package:dating/Widget/main/alarm/alarm5.dart';
-import 'package:dating/Widget/main/alarm/alarm6.dart';
-import 'package:dating/Widget/main/alarm/alarm7.dart';
-import 'package:dating/Widget/main/alarm/alarm8.dart';
-import 'package:dating/Widget/main/alarm/alarm9.dart';
+import 'package:dating/Widget/alarm/alarm1.dart';
+import 'package:dating/Widget/alarm/alarm10.dart';
+import 'package:dating/Widget/alarm/alarm2.dart';
+import 'package:dating/Widget/alarm/alarm3.dart';
+import 'package:dating/Widget/alarm/alarm4.dart';
+import 'package:dating/Widget/alarm/alarm5.dart';
+import 'package:dating/Widget/alarm/alarm6.dart';
+import 'package:dating/Widget/alarm/alarm7.dart';
+import 'package:dating/Widget/alarm/alarm8.dart';
+import 'package:dating/Widget/alarm/alarm9.dart';
 import 'package:dating/style/constant.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
