+++ lib/screen/meet/meeting_create1_screen.dart
@@ -1,5 +1,5 @@
 import 'package:dating/Widget/bottom_apply_bar.dart';
-import 'package:dating/Widget/meet/meet_create1/select_meet_method.dart';
+import 'package:dating/Widget/meet_create1/select_meet_method.dart';
 import 'package:dating/screen/meet/meeting_create2_screen.dart';
 import 'package:dating/style/constant.dart';
 import 'package:dating/style/icon_shape.dart';
