+++ lib/screen/profile/profile_edit_screen.dart
@@ -3,8 +3,8 @@ import 'package:dating/Widget/profile/ideal_type.dart';
 import 'package:dating/Widget/profile/interest.dart';
 import 'package:dating/Widget/profile/personal_information.dart';
 import 'package:dating/Widget/profile/personality.dart';
-import 'package:dating/Widget/profile/profile_edit/my_photos.dart';
-import 'package:dating/Widget/profile/profile_edit/input_field.dart';
+import 'package:dating/Widget/profile_edit/my_photos.dart';
+import 'package:dating/Widget/profile_edit/input_field.dart';
 import 'package:dating/style/icon_shape.dart';
 import 'package:dating/style/text_styling.dart';
 import 'package:flutter/material.dart';
